cool-magazine-workspace/libs/shared/ddd-expc/src/lib/implementations/specifications/uris/components/authority/user-info/no-user-info-authority-uri-component-specification.ts
import {
  CompositeSpecification,
  IUserInfoAuthorityUriComponentPart,
} from '@bh2old/ddd-expc/abstractions';

type NoUserInfoAuthorityUriComponentCandidate =
  | IUserInfoAuthorityUriComponentPart
  | string;

export class NoUserInfoAuthorityUriComponentSpecification extends CompositeSpecification<NoUserInfoAuthorityUriComponentCandidate> {
  isSatisfiedBy(candidate: NoUserInfoAuthorityUriComponentCandidate): boolean {
    const userInfo =
      typeof candidate === 'object' ? candidate.userInfo : candidate;

    return !userInfo;
  }
}
